change sequencenum type from double to numeric(15,4)

+++ study/src/org/labkey/study/model/DatasetDomainKind.java
@@ -71,7 +71,7 @@ public abstract class DatasetDomainKind extends AbstractDomainKind
         {
             new PropertyStorageSpec(PARTICIPANTID, Types.VARCHAR, 32),
             new PropertyStorageSpec(LSID, Types.VARCHAR, 200, true, false),
-            new PropertyStorageSpec(SEQUENCENUM, Types.FLOAT),
+            new PropertyStorageSpec(SEQUENCENUM, Types.NUMERIC),
             new PropertyStorageSpec(SOURCELSID, Types.VARCHAR, 200),
             new PropertyStorageSpec(_KEY, Types.VARCHAR, 200),
             new PropertyStorageSpec(QCSTATE, Types.INTEGER),
